+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0 under development
 
 - New #77: Add `ToDateTime` parameter attribute (@vjik)
+- New #78: Add `Trim`, `LeftTrim` and `RightTrim` parameter attributes (@vjik)
 - Enh #76: Raise the minimum version of PHP to 8.1 (@vjik)
 
 ## 1.1.0 February 09, 2024